공지사항 완료

+++ app/src/main/java/com/bitcamp/testproject/web/controller/BoardController.java
@@ -1,7 +1,6 @@
 package com.bitcamp.testproject.web.controller;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -19,12 +18,14 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
+import com.bitcamp.testproject.service.BoardCommentService;
 import com.bitcamp.testproject.service.BoardService;
-import com.bitcamp.testproject.vo.AttachedFile;
 import com.bitcamp.testproject.vo.Board;
 import com.bitcamp.testproject.vo.BoardCategory;
 import com.bitcamp.testproject.vo.Criteria;
+import com.bitcamp.testproject.vo.Member;
 import com.bitcamp.testproject.vo.PageMaker;
+import com.bitcamp.testproject.vo.Search;
 
 @Controller
 @RequestMapping("/board/")
@@ -32,138 +33,11 @@ public class BoardController {
 
   @Autowired
   ServletContext sc;
+
   @Autowired
   BoardService boardService;
-
-  // InternalResourceViewResolver 사용 후:
-  //  @GetMapping("form")
-  //  public void form() throws Exception {
-  //  }
-
-  //  @PostMapping("add") 
-  //  public String add(
-  //      Board board,
-  //      MultipartFile[] files,
-  //      HttpSession session) throws Exception {
-  //
-  //    board.setAttachedFiles(saveAttachedFiles(files));
-  //    board.setWriter((Member) session.getAttribute("loginMember"));
-  //
-  //    boardService.add(board);
-  //    return "redirect:list";
-  //  }
-  //
-  //  private List<BoardAttachedFile> saveAttachedFiles(Part[] files)
-  //      throws IOException, ServletException {
-  //    List<BoardAttachedFile> attachedFiles = new ArrayList<>();
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    for (Part part : files) {
-  //      if (part.getSize() == 0) {
-  //        continue;
-  //      }
-  //
-  //      String filename = UUID.randomUUID().toString();
-  //      part.write(dirPath + "/" + filename);
-  //      attachedFiles.add(new BoardAttachedFile(filename));
-  //    }
-  //    return attachedFiles;
-  //  }
-  //
-  //  private List<BoardAttachedFile> saveAttachedFiles(MultipartFile[] files)
-  //      throws IOException, ServletException {
-  //    List<BoardAttachedFile> attachedFiles = new ArrayList<>();
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    for (MultipartFile part : files) {
-  //      if (part.isEmpty()) {
-  //        continue;
-  //      }
-  //
-  //      String filename = UUID.randomUUID().toString();
-  //      part.transferTo(new File(dirPath + "/" + filename));
-  //      attachedFiles.add(new BoardAttachedFile(filename));
-  //    }
-  //    return attachedFiles;
-  //  }
-  //
-  //  @GetMapping("list")
-  //  public void list(Model model) throws Exception {
-  //    model.addAttribute("boards", boardService.list());
-  //  }
-  //
-  //  @GetMapping("detail")
-  //  public Map detail(int no) throws Exception {
-  //    Board board = boardService.get(no);
-  //    if (board == null) {
-  //      throw new Exception("해당 번호의 게시글이 없습니다!");
-  //    }
-  //
-  //    Map map = new HashMap();
-  //    map.put("board", board);
-  //    return map;
-  //  }
-  //
-  //  @PostMapping("update")
-  //  public String update(
-  //      Board board,
-  //      Part[] files,
-  //      HttpSession session) 
-  //          throws Exception {
-  //
-  //    board.setAttachedFiles(saveAttachedFiles(files));
-  //
-  //    checkOwner(board.getNo(), session);
-  //
-  //    if (!boardService.update(board)) {
-  //      throw new Exception("게시글을 변경할 수 없습니다!");
-  //    }
-  //
-  //    return "redirect:list";
-  //  }
-  //
-  //  private void checkOwner(int boardNo, HttpSession session) throws Exception {
-  //    Member loginMember = (Member) session.getAttribute("loginMember");
-  //    if (boardService.get(boardNo).getWriter().getNo() != loginMember.getNo()) {
-  //      throw new Exception("게시글 작성자가 아닙니다.");
-  //    }
-  //  }
-  //
-  //  @GetMapping("delete")
-  //  public String delete(
-  //      int no, 
-  //      HttpSession session) 
-  //          throws Exception {
-  //
-  //    checkOwner(no, session);
-  //    if (!boardService.delete(no)) {
-  //      throw new Exception("게시글을 삭제할 수 없습니다.");
-  //    }
-  //
-  //    return "redirect:list";
-  //  }
-  //
-  //  @GetMapping("fileDelete")
-  //  public String fileDelete(
-  //      int no,
-  //      HttpSession session) 
-  //          throws Exception {
-  //
-  //    BoardAttachedFile attachedFile = boardService.getAttachedFile(no); 
-  //
-  //    Member loginMember = (Member) session.getAttribute("loginMember");
-  //    Board board = boardService.get(attachedFile.getBoardNo()); 
-  //
-  //    if (board.getWriter().getNo() != loginMember.getNo()) {
-  //      throw new Exception("게시글 작성자가 아닙니다.");
-  //    }
-  //
-  //    if (!boardService.deleteAttachedFile(no)) {
-  //      throw new Exception("게시글 첨부파일을 삭제할 수 없습니다.");
-  //    }
-  //
-  //    return "redirect:detail?no=" + board.getNo();
-  //  }
+  @Autowired
+  BoardCommentService boardCommentService;
 
   @GetMapping("main")
   public Model main(Model model) {
@@ -182,71 +56,27 @@ public class BoardController {
 
   //  제동 메서드 추가 
   @PostMapping("add") 
-  public String add(
-      Board board,
-      int cateno,
-      Part file,
-      HttpSession session) throws Exception {
+  public String add(Board board, int cateno, Part file, HttpSession session) throws Exception {
 
-    // 카테고리 번호 넣기
+    // 카테고리 번호, 파일경로넣기, 등록회원 정보 넣기
     board.setCateno(cateno);
-
-    board.setAttachedFiles(saveAttachedFile(file));
-    //    board.setWriter((Member) session.getAttribute("loginMember"));
+    board.setThumbnail(saveAttachedFile(file));
+    board.setWriter((Member) session.getAttribute("loginMember"));
 
     boardService.add(board);
     return "redirect:list?no=" + cateno;
   }
 
-
-  private List<AttachedFile> saveAttachedFile(Part file)
-      throws IOException, ServletException {
-    List<AttachedFile> attachedFiles = new ArrayList<>();
+  private String saveAttachedFile(Part file) throws IOException, ServletException {
     String dirPath = sc.getRealPath("/board/files");
-
     // 첨부파일이 있다면 실행
     if (file.getSize() != 0) {
       String filename = UUID.randomUUID().toString();
       file.write(dirPath + "/" + filename);
-      attachedFiles.add(new AttachedFile(filename));
+      return filename;
     }
-
-    return attachedFiles;
+    return null;
   }
-
-
-
-  //  private String saveThumbnailFile(Part file) throws Exception {
-  //    if (file.getSize() == 0) {
-  //      return null;
-  //    }
-  //
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    String filename = UUID.randomUUID().toString();
-  //    file.write(dirPath + "/" + filename);
-  //
-  //    return filename;
-  //  }
-
-  //  
-  //  private List<AttachedFile> saveAttachedFiles(Part[] files)
-  //      throws IOException, ServletException {
-  //    List<AttachedFile> attachedFiles = new ArrayList<>();
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    for (Part part : files) {
-  //      if (part.getSize() == 0) {
-  //        continue;
-  //      }
-  //
-  //      String filename = UUID.randomUUID().toString();
-  //      part.write(dirPath + "/" + filename);
-  //      attachedFiles.add(new AttachedFile(filename));
-  //    }
-  //    return attachedFiles;
-  //  }
-  //  
 
   @GetMapping("form")
   public Model form(int cateno, Model model) throws Exception {
@@ -255,27 +85,30 @@ public class BoardController {
   }
 
   @GetMapping("detail")
-  public Model detail(int no, Model model, HttpServletRequest request, HttpServletResponse response) throws Exception {
+  public Model detail(int no, Model model, HttpServletRequest request, HttpServletResponse response, Criteria cri) throws Exception {
+
+    // 페이징하기 위한 연산 
+    PageMaker pageMaker = new PageMaker();
+    cri.setPerPageNum(5);
+    pageMaker.setCri(cri);
+    pageMaker.setTotalCount(boardCommentService.countCommentListTotal(no));
 
     // 조회수 증가     
     viewCountUp(no, request, response);
 
     // 게시글 꺼내기
     Board board = boardService.get(no);
-
-
     if (board == null) {
       throw new Exception("해당 번호의 게시글이 없습니다!");
     }
-    System.out.println(board.getViewCount());
+    model.addAttribute("board", board);
+    model.addAttribute("pageMaker", pageMaker);
 
-    System.out.println("몇번 실행되는거냐??");
-    return model.addAttribute("board", board);
+    return model;
   }
 
   // 조회수 증가 
   private void viewCountUp(int no, HttpServletRequest request, HttpServletResponse response) {
-
     Cookie oldCookie = null;
     Cookie[] cookies = request.getCookies();
     if (cookies != null) {
@@ -285,7 +118,6 @@ public class BoardController {
         }
       }
     }
-
     if (oldCookie != null) {
       if (!oldCookie.getValue().contains("[" + no + "]")) {
         boardService.increaseViews(no);
@@ -303,60 +135,64 @@ public class BoardController {
     }
   }
 
-
-  //  @GetMapping("list")
-  //  public void list(Model model, int no) throws Exception {
-  //    model.addAttribute("boards", boardService.list(no));
-  //  }
-
   @GetMapping("list")
-  public ModelAndView list(Criteria cri, int no) throws Exception {
+  public ModelAndView list(Criteria cri, int no, Search search) throws Exception {
+    ModelAndView mav;
+    if (no == 3) {
+      // 챌린지 게시판일 경우 (listOfClg.html을 실행하고 게시글 9개만 출력)
+      mav = new ModelAndView("board/listOfClg");
+      cri.setPerPageNum(6);
+    } else {
+      mav = new ModelAndView("board/list");
+      cri.setPerPageNum(5);
+    }
 
+    cri.setCatenoToPage(no); // 특정 게시판을 목록을 출력하기위한 설정
+
+    // 검색 기능인지 단지 목록 출력인지에 따라 다른 메서드 호출 
+    List<Map<String,Object>> list;
+    int boardTotalCount;
+    if (search.getKeyword() == null) {
+      list = boardService.list(cri);
+      boardTotalCount = boardService.countTotalBoard(no);
+    } else {
+      list = boardService.listWithKeyword(cri, search);
+      boardTotalCount = boardService.countTotalBoardWithSearch(no, search);
+    }
+    System.out.println("출력한 리스트 사이즈>>>" + list.size());
 
     // 페이징하기 위한 연산 
     PageMaker pageMaker = new PageMaker();
-    cri.setCatenoToPage(no); // 특정 게시판을 목록을 출력하기위한 설정
     pageMaker.setCri(cri);
-    pageMaker.setTotalCount(boardService.countBoardListTotal(no));
-
-
-    ModelAndView mav;
-    if (no == 3) {
-      // 챌린지 게시판일 경우 
-      mav = new ModelAndView("board/listOfClg");
-    } else {
-      mav = new ModelAndView("board/list");
-    }
+    pageMaker.setTotalCount(boardTotalCount);
 
     // 게시글 카테고리 번호 Board객체에 담아서 map객체에 담을 준비 
     Board catenoInBoard = new Board();
     catenoInBoard.setCateno(no);
 
-    List<Map<String,Object>> list = boardService.list(cri);
     mav.addObject("list", list);
     mav.addObject("pageMaker", pageMaker);
     mav.addObject("catenoInBoard", catenoInBoard);
+    mav.addObject("search", search);
 
     return mav;
   }
 
   @GetMapping("delete")
-  public String delete(
-      int no, 
-      HttpSession session) 
-          throws Exception {
+  public String delete(int no, HttpSession session) throws Exception {
 
-    // 돌아갈 게시판카테고리 넘버 찾기
+    // 삭제 후 돌아갈 게시판카테고리 넘버 찾기
     int cateno = boardService.get(no).getCateno();
+    // 회원 유효성 체크(삭제할 회원이 같이 회원인지)
+    checkOwner(no, session);
 
-    //    checkOwner(no, session);
     if (!boardService.delete(no)) {
       throw new Exception("게시글을 삭제할 수 없습니다.");
     }
 
-
     return "redirect:list?no=" + cateno;
   }
+
 
   @GetMapping("updateForm")
   public Model updateForm(int no, Model model) throws Exception {
@@ -377,7 +213,7 @@ public class BoardController {
       HttpSession session) 
           throws Exception {
 
-    board.setAttachedFiles(saveAttachedFile(file));
+    board.setThumbnail(saveAttachedFile(file));
 
     //      checkOwner(board.getNo(), session);
 
@@ -388,35 +224,32 @@ public class BoardController {
     return "redirect:list?no=" + cateno;
   }
 
-  // 여기부터 하자
   @GetMapping("fileDelete")
-  public String boardFileDelete(
-      int no,
-      HttpSession session) 
-          throws Exception {
+  public String boardFileDelete(int no, HttpSession session) throws Exception {
 
-    AttachedFile attachedFile = boardService.getAttachedFile(no); 
+    Member loginMember = (Member) session.getAttribute("loginMember");
+    Board board = boardService.get(no);
+    System.out.println(board);
 
-    //    Member loginMember = (Member) session.getAttribute("loginMember");
-    Board board = boardService.get(attachedFile.getObjectNo()); 
+    if (board.getWriter().getNo() != loginMember.getNo()) {
+      throw new Exception("게시글 작성자가 아닙니다.");
+    }
 
-    //    if (board.getWriter().getNo() != loginMember.getNo()) {
-    //      throw new Exception("게시글 작성자가 아닙니다.");
-    //    }
-
-    if (!boardService.deleteAttachedFile(no)) {
+    if (boardService.deleteThumbnail(no) == 0) {
       throw new Exception("게시글 첨부파일을 삭제할 수 없습니다.");
     }
 
     return "redirect:updateForm?no=" + board.getNo();
   }
 
+  private void checkOwner(int boardNo, HttpSession session) throws Exception {
+    Member loginMember = (Member) session.getAttribute("loginMember");
+    if (boardService.get(boardNo).getWriter().getNo() != loginMember.getNo()) {
+      throw new Exception("게시글 작성자가 아닙니다.");
+    }
+  }
 
-  /////// 제동 메서드 끝 
+
+
+
 }
-
-
-
-
-
-
